+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 * Add missing rebind to handler_alloc
 * Fix error handling in http server examples
 * Fix CMake scripts for MinGW
-* Better WebSocket decorator
 * Use BOOST_ASSERT
+* Better WebSocket decorator
+* Update and tidy documentation
 
 --------------------------------------------------------------------------------